+++ lib/message/flags.ts
@@ -1,0 +1,3 @@
+import { TwitchFlag } from "./flag";
+
+export type TwitchFlagList = TwitchFlag[];